Added endpoints for plan list and payment history

+++ go-api/models/pays.go
@@ -2,7 +2,7 @@ package models
 
 type Pays struct {
 	ID       int64    `json:"payment_id,omitempty"`
-	User     int64    `json:"user,omitempty"`
+	User     int      `json:"user,omitempty"`
 	Provider string   `json:"provider,omitempty"`
 	SupplyID int64    `json:"supply_id,omitempty"`
 	Amount   *float32 `json:"amount,omitempty"`
